+++ apps/commerce-streamer/src/main/java/com/loopers/application/ConsumerGroupConfig.java
@@ -1,4 +1,4 @@
-package com.loopers.config;
+package com.loopers.application;
 
 
 import org.apache.kafka.common.TopicPartition;
