+++ apps/commerce-streamer/src/main/java/com/loopers/application/ranking/RankValue.java
@@ -1,0 +1,11 @@
+package com.loopers.application.ranking;
+
+
+public record RankValue(
+        Long productId,
+        Long likeCount,
+        Long saleCount,
+        Long viewCount
+){
+
+}
