+++ modules/jpa/src/main/java/com/loopers/application/jpa/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.loopers.config.jpa;
+package com.loopers.application.jpa;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
